enable security component to work with Oracle database backend and be backward compatible with HSQL involved using BIGINT because of the way that OJB uses setObject on its JDBC objects also involved converting ints to longs, shorts to ints
PR:
Obtained from:
Submitted by:	
Reviewed by:	
CVS: ----------------------------------------------------------------------
CVS: PR:
CVS:   If this change addresses a PR in the problem report tracking
CVS:   database, then enter the PR number(s) here.
CVS: Obtained from:
CVS:   If this change has been taken from another system, such as NCSA,
CVS:   then name the system in this line, otherwise delete it.
CVS: Submitted by:
CVS:   If this code has been contributed to Apache by someone else; i.e.,
CVS:   they sent us a patch or a new module, then include their name/email
CVS:   address here. If this is your work then delete this line.
CVS: Reviewed by:
CVS:   If we are doing pre-commit code reviews and someone else has
CVS:   reviewed your changes, include their name(s) here.
CVS:   If you have not had it reviewed then delete this line.


git-svn-id: http://svn.apache.org/repos/asf/portals/jetspeed-2/trunk@186414 13f79535-47bb-0310-9956-ffa450edef68

+++ components/security/src/java/org/apache/jetspeed/security/impl/UserManagerImpl.java
@@ -67,7 +67,7 @@ public class UserManagerImpl extends BaseSecurityImpl implements UserManager
         JetspeedUserPrincipal omUser = super.getJetspeedUserPrincipal(username);
         Collection credentials = omUser.getCredentials();
         // Create a new credential with the given password.
-        short credentialType = 0;
+        int credentialType = 0;
         JetspeedCredential omCredential = new JetspeedCredentialImpl(omUser.getPrincipalId(), password, credentialType, null);
         if (log.isDebugEnabled())
             log.debug("Credential: " + omCredential.toString());
